+++ shared-components/backend/package/order/model/request.go
@@ -1,5 +1,5 @@
 package model
 
 type UpdateOrderStatusRequest struct {
-	Status string `json:"status" validate:"required,oneof=Created Paid Cancelled Fulfilled"`
+	Status string `json:"status" validate:"required"`
 }