+++ shared-components/backend/package/order/model/status.go
@@ -1,0 +1,12 @@
+package model
+
+const (
+	OrderStatusSubmitted      = "Order Submitted"
+	OrderStatusConfirmed      = "Order Confirmed"
+	OrderStatusFulfillment    = "Order Fulfillment"
+	OrderStatusOutForDelivery = "Order Out For Delivery"
+	OrderStatusDelivered      = "Order Delivered"
+	OrderStatusComplete       = "Order Complete"
+	OrderStatusReturned       = "Order Returned"
+	OrderStatusCancelled      = "Order Cancelled"
+)